:recycle: [#4] 유저 정보 조회 및 전체 구조 설계
