+++ src/main/java/gg/favorites/config/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package gg.favorites.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaAuditingConfiguration {
+}
