+++ src/main/java/gg/favorites/domain/user/UserReader.java
@@ -1,0 +1,5 @@
+package gg.favorites.domain.user;
+
+public interface UserReader {
+    User getUserByUserId(String userId);
+}
