+++ src/main/java/gg/favorites/domain/user/UserService.java
@@ -1,0 +1,10 @@
+package gg.favorites.domain.user;
+
+import gg.favorites.api.user.UserInfo;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+public interface UserService {
+
+    UserInfo findByUserId(String userId);
+}
